Modeles composants OK

+++ src/wattwatt/connectors/energyproviders/EnergyProviderConnector.java
@@ -20,7 +20,7 @@ import wattwatt.interfaces.energyproviders.IEnergyProviders;
 public abstract class EnergyProviderConnector extends AbstractConnector implements IEnergyProviders {
 
 	@Override
-	public int getEnergy() throws Exception {
+	public double getEnergy() throws Exception {
 		return ((IEnergyProviders) this.offering).getEnergy();
 	}
